# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.CreateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        minimum: 0
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  models.SubscriptionResponse:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.TotalCostRequest:
    properties:
      end_date:
        type: string
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - end_date
    - start_date
    type: object
  models.TotalCostResponse:
    properties:
      total_cost:
        type: integer
    type: object
  models.UpdateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@subscription-service.com
    name: API Support
  description: REST API for managing user subscriptions
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Subscription Service API
  version: "1.0"
paths:
  /subscriptions:
    post:
      consumes:
      - application/json
      description: Create a new subscription record
      parameters:
      - description: Subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/models.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Delete subscription by ID
      parameters:
      - description: Subscription ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      description: Get subscription details by ID
      parameters:
      - description: Subscription ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SubscriptionResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Update subscription details
      parameters:
      - description: Subscription ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/models.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update subscription
      tags:
      - subscriptions
  /subscriptions/total-cost:
    post:
      consumes:
      - application/json
      description: Calculate total cost of active subscriptions within date range
        with optional filters
      parameters:
      - description: Filter criteria
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.TotalCostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TotalCostResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get total cost of subscriptions
      tags:
      - subscriptions
schemes:
- http
swagger: "2.0"